refactor(folder): 폴더구조 변경에 따른 파일이동 및 경로 수정
Resolves: #256
