+++ features/meetupDetail/components/CommentCards/index.tsx
@@ -7,9 +7,9 @@ import { Rating, Heart } from "@smastrom/react-rating";
 import { formatIsoDateWithDots } from "@/utils/date";
 import ActionDropdown from "@/components/ui/Dropdowns/ActionDropdown";
 import { User } from "@/features/meetupDetail/types";
-import Empty from "@/components/layout/Empty";
+import Empty from "@/components/ui/Empty";
 import { useUserStore } from "@/store/user.store";
-import ReviewModal, { ReviewFormValues } from "@/components/ui/Modals/ReviewModal";
+import ReviewModal, { ReviewFormValues } from "@/features/shared/components/ReviewModal";
 import { useDeleteReviewMutation, useEditReviewMutation } from "@/features/meetupDetail/mutations";
 import Alert from "@/components/ui/Modals/AlertModal";
 import Avatar from "@/components/ui/Avatar";
