+++ features/mypage/Created/index.tsx
@@ -5,13 +5,13 @@ import { DetailCardBadge } from "@/features/mypage/types";
 import { CreatedItem } from "@/features/mypage/types";
 import Alert from "@/components/ui/Modals/AlertModal";
 import useMeetingFavorite from "@/hooks/useMeetingFavorite";
-import Empty from "@/components/layout/Empty";
+import Empty from "@/components/ui/Empty";
 import { useMyCreatedInfinite } from "../queries";
 import { useIntersectionObserver } from "@/hooks/useIntersectionObserver";
 import Loading from "@/components/ui/Loading";
 import DetailCardSkeleton from "../components/DetailCard/DetailCardSkeleton";
 import { useDeleteMeetings } from "../mutations";
-import QueryErrorBoundary from "../components/QueryErrorBoundary";
+import QueryErrorBoundary from "@/components/common/QueryErrorBoundary";
 
 // 모임 배지 상태
 function meetupBadges(item: CreatedItem): DetailCardBadge[] {
